pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ds-rpc-01"
version = "0.1.0"
description = "RAG-based assistant for FinSolve Technologies"
authors = [
    {name = "Parambrata Ghosh", email = "parambrata@finsolve.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "python-dotenv>=1.0.0",
    "pyjwt>=2.8.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.1",
    "langchain-groq>=0.0.1",
    "langchain-huggingface>=0.0.1",
    "langchain-chroma>=0.1.0",
    "chromadb>=0.4.0",
    "spacy>=3.7.0",
    "pandas>=2.0.0",
    "gunicorn>=21.2.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "pytest>=7.0.0",
]

[tool.setuptools]
packages = ["app"]
